requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.4.2
idna==2.9
mysql-connector-python==8.0.20
protobuf==3.12.2
python-dateutil==2.9.0.post0
python-dotenv==0.13.0
pytz==2020.1
requests==2.32.4
setuptools==80.9.0
six==1.15.0
tzlocal==2.1
urllib3==2.4.0APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.4.2
idna==2.9
mysql-connector-python==8.0.20
protobuf==3.12.2
python-dateutil==2.9.0.post0
python-dotenv==0.13.0
pytz==2020.1
requests==2.32.4
setuptools==80.9.0
six==1.15.0
tzlocal==2.1
urllib3==2.4.0
